client: move container options together with their users
