+++ client/container_list.go
@@ -11,6 +11,17 @@ import (
 	"github.com/moby/moby/api/types/versions"
 )
 
+// ContainerListOptions holds parameters to list containers with.
+type ContainerListOptions struct {
+	Size    bool
+	All     bool
+	Latest  bool
+	Since   string
+	Before  string
+	Limit   int
+	Filters filters.Args
+}
+
 // ContainerList returns the list of containers in the docker host.
 func (cli *Client) ContainerList(ctx context.Context, options ContainerListOptions) ([]container.Summary, error) {
 	query := url.Values{}
