+++ client/container_logs.go
@@ -10,6 +10,18 @@ import (
 	"github.com/moby/moby/client/internal/timestamp"
 )
 
+// ContainerLogsOptions holds parameters to filter logs with.
+type ContainerLogsOptions struct {
+	ShowStdout bool
+	ShowStderr bool
+	Since      string
+	Until      string
+	Timestamps bool
+	Follow     bool
+	Tail       string
+	Details    bool
+}
+
 // ContainerLogs returns the logs generated by a container in an [io.ReadCloser].
 // It's up to the caller to close the stream.
 //
